move and extend test
